+++ app/src/main/java/com/puzzling/puzzlingaos/data/datasource/local/ExampleLocalDataSource.kt
@@ -1,10 +1,10 @@
-package com.puzzling.puzzlingaos.data.source.remote
+package com.puzzling.puzzlingaos.data.datasource.local
 
 import com.puzzling.puzzlingaos.data.model.request.ExampleRequest
 import com.puzzling.puzzlingaos.data.model.response.ExampleResponse
 import com.puzzling.puzzlingaos.data.service.ExampleService
 
-class ExampleRemoteDataSource(
+class ExampleLocalDataSource(
     private val exampleService: ExampleService,
 ) {
     suspend fun postExample(exampleRequestDto: ExampleRequest): ExampleResponse =
