+++ app/src/main/java/com/puzzling/puzzlingaos/data/datasource/remote/ExampleRemoteDataSource.kt
@@ -1,10 +1,10 @@
-package com.puzzling.puzzlingaos.data.source.local
+package com.puzzling.puzzlingaos.data.datasource.remote
 
 import com.puzzling.puzzlingaos.data.model.request.ExampleRequest
 import com.puzzling.puzzlingaos.data.model.response.ExampleResponse
 import com.puzzling.puzzlingaos.data.service.ExampleService
 
-class ExampleLocalDataSource(
+class ExampleRemoteDataSource(
     private val exampleService: ExampleService,
 ) {
     suspend fun postExample(exampleRequestDto: ExampleRequest): ExampleResponse =
