+++ app/src/main/java/com/puzzling/puzzlingaos/data/datasource/remote/ProjectDataSource.kt
@@ -1,4 +1,4 @@
-package com.puzzling.puzzlingaos.data.source.remote
+package com.puzzling.puzzlingaos.data.datasource.remote
 
 import com.puzzling.puzzlingaos.data.model.request.RequestInvitationCode
 import com.puzzling.puzzlingaos.data.model.request.RequestProjectRegisterDto
