+++ app/src/main/java/com/puzzling/puzzlingaos/data/datasource/remote/WriteReviewDataSource.kt
@@ -1,4 +1,4 @@
-package com.puzzling.puzzlingaos.data.source.remote
+package com.puzzling.puzzlingaos.data.datasource.remote
 
 import com.puzzling.puzzlingaos.data.model.request.RequestReview5FDto
 import com.puzzling.puzzlingaos.data.model.request.RequestReviewAARDto
