+++ app/src/main/java/com/puzzling/puzzlingaos/data/datasource/remote/impl/MyPageDataSourceImpl.kt
@@ -1,9 +1,9 @@
-package com.puzzling.puzzlingaos.data.source.remote.impl
+package com.puzzling.puzzlingaos.data.datasource.remote.impl
 
+import com.puzzling.puzzlingaos.data.datasource.remote.MyPageDataSource
 import com.puzzling.puzzlingaos.data.model.response.ResponseDetailRetroDto
 import com.puzzling.puzzlingaos.data.model.response.ResponseMyRetroListDto
 import com.puzzling.puzzlingaos.data.service.MyPageService
-import com.puzzling.puzzlingaos.data.source.remote.MyPageDataSource
 import javax.inject.Inject
 
 class MyPageDataSourceImpl @Inject constructor(private val apiService: MyPageService) :
@@ -12,6 +12,7 @@ class MyPageDataSourceImpl @Inject constructor(private val apiService: MyPageSer
         memberId: Int,
         projectId: Int,
     ): ResponseMyRetroListDto = apiService.getMyProjectReview(memberId, projectId)
+
     override suspend fun getMyDetailReview(
         memberId: Int,
         projectId: Int,
