+++ app/src/main/java/com/puzzling/puzzlingaos/data/datasource/remote/impl/ProjectDataSourceImpl.kt
@@ -1,5 +1,6 @@
-package com.puzzling.puzzlingaos.data.source.remote.impl
+package com.puzzling.puzzlingaos.data.datasource.remote.impl
 
+import com.puzzling.puzzlingaos.data.datasource.remote.ProjectDataSource
 import com.puzzling.puzzlingaos.data.model.request.RequestInvitationCode
 import com.puzzling.puzzlingaos.data.model.request.RequestProjectRegisterDto
 import com.puzzling.puzzlingaos.data.model.response.ResponseInvitationCodeDto
@@ -7,7 +8,6 @@ import com.puzzling.puzzlingaos.data.model.response.ResponseJoinProjectDto
 import com.puzzling.puzzlingaos.data.model.response.ResponseProjectRegisterDto
 import com.puzzling.puzzlingaos.data.model.response.ResponseProjectRetroWeekDto
 import com.puzzling.puzzlingaos.data.service.ProjectService
-import com.puzzling.puzzlingaos.data.source.remote.ProjectDataSource
 import javax.inject.Inject
 
 class ProjectDataSourceImpl @Inject constructor(
