+++ app/src/main/java/com/puzzling/puzzlingaos/data/repository/ExampleRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.puzzling.puzzlingaos.data.repository
 
+import com.puzzling.puzzlingaos.data.datasource.remote.ExampleRemoteDataSource
 import com.puzzling.puzzlingaos.data.model.request.ExampleRequest
 import com.puzzling.puzzlingaos.data.model.response.ExampleResponse
-import com.puzzling.puzzlingaos.data.source.remote.ExampleRemoteDataSource
 import com.puzzling.puzzlingaos.domain.repository.ExampleRepository
 
 class ExampleRepositoryImpl(
