+++ app/src/main/java/com/puzzling/puzzlingaos/data/repository/MyBoardRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.puzzling.puzzlingaos.data.repository
 
+import com.puzzling.puzzlingaos.data.datasource.remote.MyDashBoardDataSource
 import com.puzzling.puzzlingaos.data.model.response.ResponseMyPuzzleBoardDto
-import com.puzzling.puzzlingaos.data.source.remote.MyDashBoardDataSource
 import com.puzzling.puzzlingaos.domain.entity.ActionPlan
 import com.puzzling.puzzlingaos.domain.entity.MyPuzzleBoard
 import com.puzzling.puzzlingaos.domain.entity.Project
