+++ app/src/main/java/com/puzzling/puzzlingaos/data/repository/TeamReviewRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package com.puzzling.puzzlingaos.data.repository
 
 import android.util.Log
+import com.puzzling.puzzlingaos.data.datasource.remote.TeamReviewDataSource
 import com.puzzling.puzzlingaos.data.model.response.ResponseTeamPuzzleBoardDto
 import com.puzzling.puzzlingaos.data.model.response.ResponseTeamReviewListDto
-import com.puzzling.puzzlingaos.data.source.remote.TeamReviewDataSource
 import com.puzzling.puzzlingaos.domain.entity.TeamPuzzleBoard
 import com.puzzling.puzzlingaos.domain.entity.TeamRanking
 import com.puzzling.puzzlingaos.domain.repository.TeamReviewRepository
