+++ app/src/main/java/com/puzzling/puzzlingaos/di/DataSourceModule.kt
@@ -1,15 +1,15 @@
 package com.puzzling.puzzlingaos.di
 
-import com.puzzling.puzzlingaos.data.source.remote.MyDashBoardDataSource
-import com.puzzling.puzzlingaos.data.source.remote.MyPageDataSource
-import com.puzzling.puzzlingaos.data.source.remote.ProjectDataSource
-import com.puzzling.puzzlingaos.data.source.remote.TeamReviewDataSource
-import com.puzzling.puzzlingaos.data.source.remote.WriteReviewDataSource
-import com.puzzling.puzzlingaos.data.source.remote.impl.MyDashBoardDataSourceImpl
-import com.puzzling.puzzlingaos.data.source.remote.impl.MyPageDataSourceImpl
-import com.puzzling.puzzlingaos.data.source.remote.impl.ProjectDataSourceImpl
-import com.puzzling.puzzlingaos.data.source.remote.impl.TeamReviewDataSourceImpl
-import com.puzzling.puzzlingaos.data.source.remote.impl.WriteReviewDataSourceImpl
+import com.puzzling.puzzlingaos.data.datasource.remote.MyDashBoardDataSource
+import com.puzzling.puzzlingaos.data.datasource.remote.MyPageDataSource
+import com.puzzling.puzzlingaos.data.datasource.remote.ProjectDataSource
+import com.puzzling.puzzlingaos.data.datasource.remote.TeamReviewDataSource
+import com.puzzling.puzzlingaos.data.datasource.remote.WriteReviewDataSource
+import com.puzzling.puzzlingaos.data.datasource.remote.impl.MyDashBoardDataSourceImpl
+import com.puzzling.puzzlingaos.data.datasource.remote.impl.MyPageDataSourceImpl
+import com.puzzling.puzzlingaos.data.datasource.remote.impl.ProjectDataSourceImpl
+import com.puzzling.puzzlingaos.data.datasource.remote.impl.TeamReviewDataSourceImpl
+import com.puzzling.puzzlingaos.data.datasource.remote.impl.WriteReviewDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
